msm: mdss: Power on display asynchronously as early as possible
Currently, mdss powers on the display a long time after an unblank is
requested and gets completely blocked while powering on the display, so if
the display takes a very long time to turn on, then mdss will be stuck
unable to do anything else in the meantime. This results in a long delay
between trying to wake the device and the display actually powering on
and rendering a frame.

In order to make the display turn on faster when waking the device from
sleep, start powering on the display as soon as the framebuffer unblank
event is received. This allows mdss to continue resuming while the display
takes its time powering on. A high-priority kthread is used here to ensure
the display powers on as quickly as possible.

To make this work, kickoffs need to be blocked when they attempt to power
on the display, so that a kickoff won't continue while the display is
still powered off.

Signed-off-by: Sultan Alsawaf <sultanxda@gmail.com>

# Conflicts:
#	drivers/video/fbdev/msm/mdss_dsi.c
#	drivers/video/fbdev/msm/mdss_dsi.h