Revert "cfg80211: add warning when calculating MCS rates >= 32"
Unfortunately, hdd_add_tx_bitrate in qcacld-3.0 calls this with a
clearly greater than 32 value because it spams dmesg. Right now, it's
not possible to build the driver into the kernel image and there is no
clean place to put a new module (because it's on /vendor and this is a
system-as-root device) so we'll just silence the warning for now.

Hopefully come P, the Wi-Fi HAL can handle both a module and a built-in
driver rather than one or the other.

This reverts commit 2615f3759b7b7d47a2fa894bc9f45286bdf122d7.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>