Revert "msm: camera: Fix power related issues across sensor modules"
With this commit, the camera sensor is completely broken (nothing loads
in the camera app and Face Unlock does not work).

Because this is a non-trivial change, it will take a while to debug
which part results in the breakage (since there are a few different
things happening in this change) and then address it. In order to get
the rest of LA.UM.6.3.r4-05000-sdm845.0 out in a timely manner, I am
reverting the entire change and will revisit it later. It may also turn
out to be completely outside of our control if it requires a newer
camera HAL or blobs.

This reverts commit 3b7805c1f2b225a030ec052acb7f25efa0fa4e03.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>