staging: lowmemorykiller: Checkout to android-4.9
Qualcomm adds all of this junk that isn't necessary for what is
supposed to be a simple memory management driver. I wanted to
switch to the userspace LMK daemon (conveniently named lmkd) but
unfortunately, OnePlus removed the '/dev/memcg' mounts from init.rc,
meaning Magisk would be required to overlay the ramdisk files that
would create and configure that mount. This should be the next best
thing.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>